+++ src/transform/TransformSequence.hpp
@@ -143,7 +143,7 @@ namespace kanzi {
            if ((output._index + count > output._length) || (in->_index + count > in->_length))
                _skipFlags = SKIP_MASK;
            else
-               std::memcpy(&output._array[output._index], &in->_array[in->_index], count);
+               std::memcpy(&output._array[output._index], &in->_array[in->_index], size_t(count));
        }
 
        input._index += blockSize;
@@ -167,7 +167,7 @@ namespace kanzi {
            return true;
 
        if (_skipFlags == SKIP_MASK) {
-           std::memcpy(&output._array[output._index], &input._array[input._index], count);
+           std::memcpy(&output._array[output._index], &input._array[input._index], size_t(count));
            input._index += count;
            output._index += count;
            return true;
@@ -215,7 +215,7 @@ namespace kanzi {
            if ((output._index + count > output._length) || (input._index + count > input._length))
                res = false;
            else
-               std::memcpy(&output._array[output._index], &input._array[input._index], count);
+               std::memcpy(&output._array[output._index], &input._array[input._index], size_t(count));
        }
 
        input._index += blockSize;
